requirements.txt
Cython==0.29.33
python-magic==0.4.27
python-magic-bin==0.4.14
toml
